fix: MainNavigator 사용하지 않고 ViewModel에서 처리하도록 변경
- MainActivity.kt: RouteStack을 통한 스택 관리 로직 추가
- MainViewModel.kt: 네비게이션 관련 상태 관리 메서드 추가
- MainScreen.kt: RouteStack을 활용한 UI 업데이트
