+++ core/navigation/src/main/java/com/teambrake/brake/core/navigation/route/Route.kt
@@ -1,6 +1,5 @@
 package com.teambrake.brake.core.navigation.route
 
 import androidx.navigation3.runtime.NavKey
-import timber.log.Timber
 
-interface Route: NavKey
+interface Route : NavKey
