+++ presentation/main/src/main/java/com/teambrake/brake/presentation/main/MainActivity.kt
@@ -17,7 +17,6 @@ import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableLongStateOf
 import androidx.compose.runtime.remember
@@ -27,6 +26,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.window.Popup
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.amplitude.android.Amplitude
 import com.teambrake.brake.core.amplitude.AmplitudeEventHelper
 import com.teambrake.brake.core.amplitude.TriggerSource
@@ -40,8 +40,6 @@ import com.teambrake.brake.core.navigation.compositionlocal.LocalNavigatorProvid
 import com.teambrake.brake.presentation.main.component.BrakeSnackbarHostState
 import com.teambrake.brake.presentation.main.component.BrakeSnackbarType
 import com.teambrake.brake.presentation.main.component.LogoutWarningDialog
-import com.teambrake.brake.presentation.main.navigation.MainNavigator
-import com.teambrake.brake.presentation.main.navigation.rememberMainNavigator
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
@@ -57,6 +55,7 @@ class MainActivity : ComponentActivity() {
 
 	override fun onCreate(savedInstanceState: Bundle?) {
 		super.onCreate(savedInstanceState)
+		enableEdgeToEdge()
 
 		// 앱 시작 이벤트 트래킹
 		trackAmplitudeOpenAppEvent(intent)
@@ -68,107 +67,102 @@ class MainActivity : ComponentActivity() {
 		// 스플래시 스크린 설치, 내부에서 API 31 미만 버전도 호환되도록 처리
 		val splashScreen = installSplashScreen()
 
-		enableEdgeToEdge()
-
-		// 스플래시 스크린이 유지되는 조건 설정
-		splashScreen.setKeepOnScreenCondition {
-			viewModel.startRoute.value == null
-		}
-
-		viewModel.decideStartDestination(context = this@MainActivity)
-
 		setContent {
-			val startDestination by viewModel.startRoute.collectAsState()
+			val routeStack by viewModel.routeStack.collectAsStateWithLifecycle()
 
-			when (val destination = startDestination) {
-				null -> { /* 스플래시 화면 유지 */ }
-				else -> {
-					val navigator: MainNavigator = rememberMainNavigator(destination)
-					val coroutineScope: CoroutineScope = rememberCoroutineScope()
-					val snackBarHostState = remember { BrakeSnackbarHostState() }
+			// 스플래시 스크린이 유지되는 조건 설정
+			splashScreen.setKeepOnScreenCondition {
+				routeStack.backStack.isEmpty()
+			}
 
-					val mainAction = object : MainAction {
-						@Composable
-						override fun OnFinishBackHandler() {
-							var backPressedTime by remember { mutableLongStateOf(0L) }
-							BackHandler {
-								if (System.currentTimeMillis() - backPressedTime <= 2000L) {
-									finish()
-								} else {
-									Toast.makeText(
-										this@MainActivity,
-										this@MainActivity.getString(
-											R.string.exit_message,
-										),
-										Toast.LENGTH_SHORT,
-									).show()
-								}
-								backPressedTime = System.currentTimeMillis()
-							}
+			viewModel.decideStartDestination(context = this@MainActivity)
+
+			if (routeStack.backStack.isEmpty()) return@setContent
+
+			val coroutineScope: CoroutineScope = rememberCoroutineScope()
+			val snackBarHostState = remember { BrakeSnackbarHostState() }
+
+			val mainAction = object : MainAction {
+				@Composable
+				override fun OnFinishBackHandler() {
+					var backPressedTime by remember { mutableLongStateOf(0L) }
+					BackHandler {
+						if (System.currentTimeMillis() - backPressedTime <= 2000L) {
+							finish()
+						} else {
+							Toast.makeText(
+								this@MainActivity,
+								this@MainActivity.getString(
+									R.string.exit_message,
+								),
+								Toast.LENGTH_SHORT,
+							).show()
 						}
+						backPressedTime = System.currentTimeMillis()
+					}
+				}
 
-						@Composable
-						override fun OnShowLogoutDialog(
-							onConfirm: () -> Unit,
-							onDismiss: () -> Unit,
+				@Composable
+				override fun OnShowLogoutDialog(
+					onConfirm: () -> Unit,
+					onDismiss: () -> Unit,
+				) {
+					LogoutWarningDialog(
+						onConfirm = onConfirm,
+						onDismissRequest = onDismiss,
+					)
+				}
+
+				@Composable
+				override fun OnShowLoading() {
+					Popup {
+						Box(
+							modifier = Modifier
+								.fillMaxSize()
+								.background(MaterialTheme.colorScheme.background.copy(alpha = 0.8f))
+								.statusBarsPadding(),
+							contentAlignment = Alignment.Center,
 						) {
-							LogoutWarningDialog(
-								onConfirm = onConfirm,
-								onDismissRequest = onDismiss,
-							)
-						}
-
-						@Composable
-						override fun OnShowLoading() {
-							Popup {
-								Box(
-									modifier = Modifier
-										.fillMaxSize()
-										.background(MaterialTheme.colorScheme.background.copy(alpha = 0.8f))
-										.statusBarsPadding(),
-									contentAlignment = Alignment.Center,
-								) {
-									DotProgressIndicator()
-								}
-							}
-						}
-
-						override fun onShowErrorMessage(message: String) {
-							coroutineScope.launch {
-								snackBarHostState.showSnackbar(
-									message = message,
-									actionLabel = BrakeSnackbarType.ERROR.name,
-									duration = 5000L,
-									onAction = snackBarHostState::dismiss,
-								)
-							}
-						}
-
-						override fun onShowSuccessMessage(message: String) {
-							coroutineScope.launch {
-								snackBarHostState.showSnackbar(
-									message = message,
-									actionLabel = BrakeSnackbarType.SUCCESS.name,
-									duration = 3000L,
-									onAction = snackBarHostState::dismiss,
-								)
-							}
+							DotProgressIndicator()
 						}
 					}
+				}
 
-					CompositionLocalProvider(
-						LocalMainAction provides mainAction,
-						LocalNavigatorAction provides navigator.navigatorAction(),
-						LocalNavigatorProvider provides navigator.navigatorProvider(),
-					) {
-						BrakeTheme {
-							MainScreen(
-								navigator = navigator,
-								onChangeDarkTheme = { false },
-								snackBarHostState = snackBarHostState,
-							)
-						}
+				override fun onShowErrorMessage(message: String) {
+					coroutineScope.launch {
+						snackBarHostState.showSnackbar(
+							message = message,
+							actionLabel = BrakeSnackbarType.ERROR.name,
+							duration = 5000L,
+							onAction = snackBarHostState::dismiss,
+						)
 					}
+				}
+
+				override fun onShowSuccessMessage(message: String) {
+					coroutineScope.launch {
+						snackBarHostState.showSnackbar(
+							message = message,
+							actionLabel = BrakeSnackbarType.SUCCESS.name,
+							duration = 3000L,
+							onAction = snackBarHostState::dismiss,
+						)
+					}
+				}
+			}
+
+			CompositionLocalProvider(
+				LocalMainAction provides mainAction,
+				LocalNavigatorAction provides viewModel.navigatorAction(),
+				LocalNavigatorProvider provides viewModel.navigatorProvider(),
+			) {
+				BrakeTheme {
+					MainScreen(
+						routeStack = routeStack,
+						onTabSelected = viewModel::navigateTab,
+						onChangeDarkTheme = { false },
+						snackBarHostState = snackBarHostState,
+					)
 				}
 			}
 		}
