+++ presentation/main/src/main/java/com/teambrake/brake/presentation/main/MainScreen.kt
@@ -13,7 +13,6 @@ import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -21,28 +20,29 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.teambrake.brake.core.designsystem.theme.DynamicPaddingsProvider
 import com.teambrake.brake.core.designsystem.theme.LocalDynamicPaddings
 import com.teambrake.brake.core.navigation.route.InitialRoute
 import com.teambrake.brake.core.navigation.route.MainTabRoute
+import com.teambrake.brake.core.navigation.route.RouteStack
 import com.teambrake.brake.presentation.main.component.BrakeSnackbar
-import com.teambrake.brake.presentation.main.component.BrakeSnackbarHostState
 import com.teambrake.brake.presentation.main.component.BrakeSnackbarHost
+import com.teambrake.brake.presentation.main.component.BrakeSnackbarHostState
 import com.teambrake.brake.presentation.main.navigation.MainBottomNavBar
 import com.teambrake.brake.presentation.main.navigation.MainNavHost
-import com.teambrake.brake.presentation.main.navigation.MainNavigator
 import com.teambrake.brake.presentation.main.navigation.MainTab
 import kotlinx.collections.immutable.toPersistentList
 
 @Composable
 internal fun MainScreen(
-	navigator: MainNavigator,
+	routeStack: RouteStack,
+	onTabSelected: (MainTab) -> Unit,
 	onChangeDarkTheme: (Boolean) -> Unit,
 	snackBarHostState: BrakeSnackbarHostState,
 ) {
 	MainScreenContent(
-		navigator = navigator,
+		routeStack = routeStack,
+		onTabSelected = onTabSelected,
 		onChangeDarkTheme = onChangeDarkTheme,
 		snackBarHostState = snackBarHostState,
 	)
@@ -50,15 +50,12 @@ internal fun MainScreen(
 
 @Composable
 private fun MainScreenContent(
-	navigator: MainNavigator,
+	routeStack: RouteStack,
+	onTabSelected: (MainTab) -> Unit,
 	onChangeDarkTheme: (Boolean) -> Unit,
 	snackBarHostState: BrakeSnackbarHostState,
 	modifier: Modifier = Modifier,
 ) {
-	// 1. 현재 화면에 따라 실시간 스낵바 위치 조정을 위한 Route 구독
-	// 2. MainTabRoute 화면일 때 하단 네비게이션 바를 띄우고, 그 외는 안띄우기 위한 Route 구독
-	val routeStack by navigator.routeStack.collectAsStateWithLifecycle()
-
 	// 바텀 패딩 조정 용도 (스낵바 높이 위치 및 하단 네비게이션 바 상호작용)
 	val dynamicPaddingsProvider = remember { DynamicPaddingsProvider() }
 	val density = LocalDensity.current
@@ -73,7 +70,7 @@ private fun MainScreenContent(
 					LocalDynamicPaddings provides dynamicPaddingsProvider,
 				) {
 					MainNavHost(
-						backStack = routeStack.stack,
+						backStack = routeStack.backStack,
 						padding = padding,
 						onChangeDarkTheme = onChangeDarkTheme,
 					)
@@ -110,7 +107,7 @@ private fun MainScreenContent(
 								is MainTabRoute.Report -> MainTab.REPORT
 								is MainTabRoute.Setting -> MainTab.SETTING
 							},
-							onTabSelected = navigator::navigateTab,
+							onTabSelected = onTabSelected,
 						)
 					}
 				}
